requirements.txt
aiohttp==3.11.7
colorlog==6.9.0
lxml==5.3.0
pika==1.3.2
python-dotenv==1.0.1
